split build_evt into sub modules
